Add __str__ to PuoriborState
