+++ examples/puoribor.py
@@ -3,11 +3,11 @@ import sys
 sys.path.append("../")
 
 import os
-import re
 
 import numpy as np
-from colorama import Fore, Style, init
 from numpy.typing import NDArray
+
+import colorama
 
 from fights.base import BaseAgent
 from fights.envs import puoribor
@@ -39,76 +39,24 @@ class PuoriborAgent(BaseAgent):
         return self._rng.choice(actions)
 
 
-def fallback_to_ascii(s: str) -> str:
-    try:
-        s.encode(sys.stdout.encoding)
-    except UnicodeEncodeError:
-        s = re.sub("[┌┬┐├┼┤└┴┘]", "+", s.replace("─", "-").replace("│", "|"))
-    return s
-
-
-def get_printable_board(board: NDArray[np.int_]) -> str:
-    table_top = fallback_to_ascii("┌───┬───┬───┬───┬───┬───┬───┬───┬───┐")
-    vertical_wall = fallback_to_ascii("│")
-    horizontal_wall = fallback_to_ascii("───")
-    left_intersection = fallback_to_ascii("├")
-    middle_intersection = fallback_to_ascii("┼")
-    right_intersection = fallback_to_ascii("┤")
-    left_intersection_bottom = fallback_to_ascii("└")
-    middle_intersection_bottom = fallback_to_ascii("┴")
-    right_intersection_bottom = fallback_to_ascii("┘")
-    result = table_top + "\n"
-
-    for y in range(9):
-        board_line = board[:, :, y]
-        result += vertical_wall
-        for x in range(9):
-            board_cell = board_line[:, x]
-            if board_cell[0]:
-                result += " 0 "
-            elif board_cell[1]:
-                result += " 1 "
-            else:
-                result += "   "
-            if board_cell[3]:
-                result += Fore.RED + vertical_wall + Style.RESET_ALL
-            else:
-                result += vertical_wall
-            if x == 8:
-                result += "\n"
-        result += left_intersection_bottom if y == 8 else left_intersection
-        for x in range(9):
-            board_cell = board_line[:, x]
-            if board_cell[2]:
-                result += Fore.BLUE + horizontal_wall + Style.RESET_ALL
-            else:
-                result += horizontal_wall
-            if x == 8:
-                result += right_intersection_bottom if y == 8 else right_intersection
-            else:
-                result += middle_intersection_bottom if y == 8 else middle_intersection
-        result += "\n"
-
-    return result
-
-
 if __name__ == "__main__":
     assert puoribor.PuoriborEnv.env_id == PuoriborAgent.env_id
-
-    init()
+    colorama.init()
     state = puoribor.PuoriborEnv().initialize_state()
     agents = [PuoriborAgent(0), PuoriborAgent(1)]
-    print(get_printable_board(state.board), end="")
+
+    print("\x1b[2J")
 
     it = 0
     while not state.done:
-        os.system("cls" if os.name == "nt" else "clear")
-        print(get_printable_board(state.board), end="")
+        print("\x1b[1;1H")
+        print(state)
         for agent in agents:
             action = agent(state)
             state = puoribor.PuoriborEnv().step(state, agent.agent_id, action)
-            os.system("cls" if os.name == "nt" else "clear")
-            print(get_printable_board(state.board), end="")
+
+            print("\x1b[1;1H")
+            print(state)
             if state.done:
                 print(f"agent {agent.agent_id} won in {it} iters")
                 break
